fix: constants in APIClient
