+++ @types/ReactNative.ts
@@ -6,6 +6,5 @@ declare module "react-native" {
     export interface NativeModulesStatic {
         APIClient: NativeAPIClient;
         GenericClient: NativeGenericClient;
-        NetworkConstants: NativeConstants;
     }
 }
